fix(middleware-retry): circular dependencies
$ madge --warning --circular --extensions js dist/cjs/
Processed 12 files (426ms)

✖ Found 2 circular dependencies!

1) AdaptiveRetryStrategy.js > StandardRetryStrategy.js > configurations.js
2) StandardRetryStrategy.js > configurations.js
