+++ packages/middleware-retry/src/StandardRetryStrategy.spec.ts
@@ -2,7 +2,7 @@ import { HttpRequest } from "@aws-sdk/protocol-http";
 import { isThrottlingError } from "@aws-sdk/service-error-classification";
 import { v4 } from "uuid";
 
-import { DEFAULT_MAX_ATTEMPTS, RETRY_MODES } from "./configurations";
+import { DEFAULT_MAX_ATTEMPTS, RETRY_MODES } from "./config";
 import { DEFAULT_RETRY_DELAY_BASE, INITIAL_RETRY_TOKENS, THROTTLING_RETRY_DELAY_BASE } from "./constants";
 import { getDefaultRetryQuota } from "./defaultRetryQuota";
 import { defaultDelayDecider } from "./delayDecider";
@@ -85,7 +85,7 @@ describe("defaultStrategy", () => {
     (defaultDelayDecider as jest.Mock).mockReturnValue(0);
     (defaultRetryDecider as jest.Mock).mockReturnValue(true);
     (getDefaultRetryQuota as jest.Mock).mockReturnValue(mockDefaultRetryQuota);
-    (HttpRequest as unknown as jest.Mock).mockReturnValue({
+    ((HttpRequest as unknown) as jest.Mock).mockReturnValue({
       isInstance: jest.fn().mockReturnValue(false),
     });
     (v4 as jest.Mock).mockReturnValue("42");
@@ -234,8 +234,8 @@ describe("defaultStrategy", () => {
 
       expect(defaultDelayDecider as jest.Mock).toHaveBeenCalledTimes(maxAttempts - 1);
       expect(setTimeout).toHaveBeenCalledTimes(maxAttempts - 1);
-      expect((setTimeout as unknown as jest.Mock).mock.calls[0][1]).toBe(FIRST_DELAY);
-      expect((setTimeout as unknown as jest.Mock).mock.calls[1][1]).toBe(SECOND_DELAY);
+      expect(((setTimeout as unknown) as jest.Mock).mock.calls[0][1]).toBe(FIRST_DELAY);
+      expect(((setTimeout as unknown) as jest.Mock).mock.calls[1][1]).toBe(SECOND_DELAY);
     });
   });
 
@@ -405,7 +405,7 @@ describe("defaultStrategy", () => {
 
     it("uses a unique header for every SDK operation invocation", async () => {
       const { isInstance } = HttpRequest;
-      (isInstance as unknown as jest.Mock).mockReturnValue(true);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(true);
 
       const uuidForInvocationOne = "uuid-invocation-1";
       const uuidForInvocationTwo = "uuid-invocation-2";
@@ -424,12 +424,12 @@ describe("defaultStrategy", () => {
       expect(next.mock.calls[0][0].request.headers["amz-sdk-invocation-id"]).toBe(uuidForInvocationOne);
       expect(next.mock.calls[1][0].request.headers["amz-sdk-invocation-id"]).toBe(uuidForInvocationTwo);
 
-      (isInstance as unknown as jest.Mock).mockReturnValue(false);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(false);
     });
 
     it("uses same value for additional HTTP requests associated with an SDK operation", async () => {
       const { isInstance } = HttpRequest;
-      (isInstance as unknown as jest.Mock).mockReturnValueOnce(true);
+      ((isInstance as unknown) as jest.Mock).mockReturnValueOnce(true);
 
       const uuidForInvocation = "uuid-invocation-1";
       (v4 as jest.Mock).mockReturnValueOnce(uuidForInvocation);
@@ -440,7 +440,7 @@ describe("defaultStrategy", () => {
       expect(next.mock.calls[0][0].request.headers["amz-sdk-invocation-id"]).toBe(uuidForInvocation);
       expect(next.mock.calls[1][0].request.headers["amz-sdk-invocation-id"]).toBe(uuidForInvocation);
 
-      (isInstance as unknown as jest.Mock).mockReturnValue(false);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(false);
     });
   });
 
@@ -471,7 +471,7 @@ describe("defaultStrategy", () => {
 
     it("adds header for each attempt", async () => {
       const { isInstance } = HttpRequest;
-      (isInstance as unknown as jest.Mock).mockReturnValue(true);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(true);
 
       const mockError = new Error("mockError");
       next = jest.fn((args) => {
@@ -491,14 +491,14 @@ describe("defaultStrategy", () => {
       }
 
       expect(next).toHaveBeenCalledTimes(maxAttempts);
-      (isInstance as unknown as jest.Mock).mockReturnValue(false);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(false);
     });
   });
 
   describe("defaults maxAttempts to DEFAULT_MAX_ATTEMPTS", () => {
     it("when maxAttemptsProvider throws error", async () => {
       const { isInstance } = HttpRequest;
-      (isInstance as unknown as jest.Mock).mockReturnValue(true);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(true);
 
       next = jest.fn((args) => {
         expect(args.request.headers["amz-sdk-request"]).toBe(`attempt=1; max=${DEFAULT_MAX_ATTEMPTS}`);
@@ -512,7 +512,7 @@ describe("defaultStrategy", () => {
       await retryStrategy.retry(next, { request: { headers: {} } } as any);
 
       expect(next).toHaveBeenCalledTimes(1);
-      (isInstance as unknown as jest.Mock).mockReturnValue(false);
+      ((isInstance as unknown) as jest.Mock).mockReturnValue(false);
     });
   });
 });
