+++ packages/middleware-retry/src/index.ts
@@ -2,6 +2,7 @@ export * from "./retryMiddleware";
 export * from "./omitRetryHeadersMiddleware";
 export * from "./StandardRetryStrategy";
 export * from "./AdaptiveRetryStrategy";
+export * from "./config";
 export * from "./configurations";
 export * from "./delayDecider";
 export * from "./retryDecider";
